Added stuff to CPP version

+++ usr/src/curl.sh
@@ -2,10 +2,18 @@
 
 clear
 
-rm -r ~/.android-toolkit/
+if [[ -d ~/.android-toolkit ]]; then
+
+    rm -r ~/.android-toolkit/
+
+fi
 
 git clone https://github.com/AtlasJ2301/android-toolkit ~/.android-toolkit/
 
-sudo rm /usr/bin/android-toolkit
+if [[ -f /usr/bin/android-toolkit ]]; then
+
+    sudo rm /usr/bin/android-toolkit
+
+fi
 
 sudo git clone https://github.com/AtlasJ2301/android-toolkit/blob/main/usr/bin/android-toolkit /usr/bin/android-toolkit
